+++ ByteDance/src/main/java/com/applovin/mediation/adapters/ByteDanceMediationAdapter.java
@@ -62,6 +62,8 @@ import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 
+import androidx.annotation.Nullable;
+
 /**
  * Created by Thomas So on April 14 2020
  */
@@ -151,7 +153,7 @@ public class ByteDanceMediationAdapter
             status = InitializationStatus.INITIALIZING;
             final Bundle serverParameters = parameters.getServerParameters();
             final String appId = serverParameters.getString( "app_id" );
-            log( "Initializing ByteDance SDK with app id: " + appId + "..." );
+            log( "Initializing SDK with app id: " + appId + "..." );
 
             TTAdConfig.Builder builder = new TTAdConfig.Builder();
 
@@ -206,15 +208,12 @@ public class ByteDanceMediationAdapter
                     .supportMultiProcess( false )
                     .build();
 
-            // NOTE: `activity` can only be null in 11.1.0+, and `getApplicationContext()` is introduced in 11.1.0
-            Context context = ( activity != null ) ? activity.getApplicationContext() : getApplicationContext();
-
-            TTAdSdk.init( context, adConfig, new TTAdSdk.InitCallback()
+            TTAdSdk.init( getContext( activity ), adConfig, new TTAdSdk.InitCallback()
             {
                 @Override
                 public void success()
                 {
-                    log( "Pangle SDK initialized" );
+                    log( "SDK initialized" );
 
                     status = InitializationStatus.INITIALIZED_SUCCESS;
                     onCompletionListener.onCompletion( status, null );
@@ -223,7 +222,7 @@ public class ByteDanceMediationAdapter
                 @Override
                 public void fail(int code, String msg)
                 {
-                    log( "Pangle SDK failed to initialize with code: " + code + " and message: " + msg );
+                    log( "SDK failed to initialize with code: " + code + " and message: " + msg );
 
                     status = InitializationStatus.INITIALIZED_FAILURE;
                     onCompletionListener.onCompletion( status, msg );
@@ -232,8 +231,7 @@ public class ByteDanceMediationAdapter
         }
         else
         {
-            log( "Pangle attempted initialization already - marking initialization as completed" );
-
+            log( "attempted initialization already - marking initialization as completed" );
             onCompletionListener.onCompletion( status, null );
         }
     }
@@ -305,7 +303,7 @@ public class ByteDanceMediationAdapter
         }
 
         interstitialAdListener = new InterstitialAdListener( codeId, listener );
-        TTAdSdk.getAdManager().createAdNative( activity ).loadFullScreenVideoAd( adSlotBuilder.build(), interstitialAdListener );
+        TTAdSdk.getAdManager().createAdNative( getContext( activity ) ).loadFullScreenVideoAd( adSlotBuilder.build(), interstitialAdListener );
     }
 
     @Override
@@ -345,7 +343,7 @@ public class ByteDanceMediationAdapter
         }
 
         rewardedAdListener = new RewardedAdListener( codeId, listener );
-        TTAdSdk.getAdManager().createAdNative( activity ).loadRewardVideoAd( adSlotBuilder.build(), rewardedAdListener );
+        TTAdSdk.getAdManager().createAdNative( getContext( activity ) ).loadRewardVideoAd( adSlotBuilder.build(), rewardedAdListener );
     }
 
     @Override
@@ -385,7 +383,7 @@ public class ByteDanceMediationAdapter
             adSlotBuilder.withBid( bidResponse );
         }
 
-        TTAdNative adViewAd = TTAdSdk.getAdManager().createAdNative( activity );
+        TTAdNative adViewAd = TTAdSdk.getAdManager().createAdNative( getContext( activity ) );
         if ( isNative )
         {
             NativeAdViewListener nativeListener = new NativeAdViewListener( parameters, adFormat, activity, listener );
@@ -410,6 +408,14 @@ public class ByteDanceMediationAdapter
         String codeId = parameters.getThirdPartyAdPlacementId();
         log( "Loading " + ( isBiddingAd ? "bidding " : "" ) + "native ad for code id \"" + codeId + "\"..." );
 
+        // Minimum supported Android SDK version is 11.1.0+, previous version has `MaxNativeAdView` requiring an Activity context which might leak
+        if ( AppLovinSdk.VERSION_CODE < 11010000 )
+        {
+            log( "Failing ad load for AppLovin SDK < 11.1.0 which requires an Activity context" );
+            listener.onNativeAdLoadFailed( MaxAdapterError.UNSPECIFIED );
+            return;
+        }
+
         AdSlot.Builder adSlotBuilder = new AdSlot.Builder()
                 .setCodeId( codeId )
                 .setImageAcceptedSize( 640, 320 )
@@ -421,8 +427,8 @@ public class ByteDanceMediationAdapter
             adSlotBuilder.withBid( bidResponse );
         }
 
-        nativeAdListener = new NativeAdListener( parameters, activity, listener );
-        TTAdSdk.getAdManager().createAdNative( activity ).loadFeedAd( adSlotBuilder.build(), nativeAdListener );
+        nativeAdListener = new NativeAdListener( parameters, getContext( activity ), listener );
+        TTAdSdk.getAdManager().createAdNative( getContext( activity ) ).loadFeedAd( adSlotBuilder.build(), nativeAdListener );
     }
 
     //endregion
@@ -535,6 +541,12 @@ public class ByteDanceMediationAdapter
         }
 
         return new MaxAdapterError( adapterError.getErrorCode(), adapterError.getErrorMessage(), byteDanceErrorCode, byteDanceErrorMessage );
+    }
+
+    private Context getContext(@Nullable Activity activity)
+    {
+        // NOTE: `activity` can only be null in 11.1.0+, and `getApplicationContext()` is introduced in 11.1.0
+        return ( activity != null ) ? activity.getApplicationContext() : getApplicationContext();
     }
 
     //endregion
@@ -849,7 +861,9 @@ public class ByteDanceMediationAdapter
                         log( "Adding native ad icon (" + nativeAdViewAd.getIcon().getImageUrl() + ") to queue to be fetched" );
 
                         final String imageUrl = nativeAdViewAd.getIcon().getImageUrl();
-                        iconDrawableFuture = ( AppLovinSdk.VERSION_CODE >= 11000000 ) ? createDrawableFuture( imageUrl, resources ) : executorServiceToUse.submit( createDrawableTask( imageUrl, resources ) );
+                        iconDrawableFuture = ( AppLovinSdk.VERSION_CODE >= 11000000 )
+                                ? createDrawableFuture( imageUrl, resources )
+                                : executorServiceToUse.submit( createDrawableTask( imageUrl, resources ) );
                     }
 
                     // Pangle's media view can be either a video or image (which they don't provide a view for)
@@ -867,7 +881,9 @@ public class ByteDanceMediationAdapter
                             log( "Adding native ad media (" + ttMediaImage.getImageUrl() + ") to queue to be fetched" );
 
                             final String imageUrl = ttMediaImage.getImageUrl();
-                            imageDrawableFuture = ( AppLovinSdk.VERSION_CODE >= 11000000 ) ? createDrawableFuture( imageUrl, resources ) : executorServiceToUse.submit( createDrawableTask( imageUrl, resources ) );
+                            imageDrawableFuture = ( AppLovinSdk.VERSION_CODE >= 11000000 )
+                                    ? createDrawableFuture( imageUrl, resources )
+                                    : executorServiceToUse.submit( createDrawableTask( imageUrl, resources ) );
                         }
                     }
 
@@ -1055,31 +1071,22 @@ public class ByteDanceMediationAdapter
     {
         final String                     codeId;
         final Bundle                     serverParameters;
-        final WeakReference<Activity>    activityRef;
+        final Context                    context;
         final MaxNativeAdAdapterListener listener;
 
         NativeAdListener(final MaxAdapterResponseParameters parameters,
-                         final Activity activity,
+                         final Context context,
                          final MaxNativeAdAdapterListener listener)
         {
             this.codeId = parameters.getThirdPartyAdPlacementId();
             this.serverParameters = parameters.getServerParameters();
-            this.activityRef = new WeakReference<>( activity );
+            this.context = context;
             this.listener = listener;
         }
 
         @Override
         public void onFeedAdLoad(final List<TTFeedAd> ads)
         {
-            final Activity activity = activityRef.get();
-            if ( activity == null )
-            {
-                log( "Native ad (" + codeId + ") failed to load: activity reference is null when ad is loaded" );
-                listener.onNativeAdLoadFailed( MaxAdapterError.INVALID_LOAD_STATE );
-
-                return;
-            }
-
             if ( ads == null || ads.size() == 0 )
             {
                 log( "Native ad (" + codeId + ") failed to load: no fill" );
@@ -1115,7 +1122,7 @@ public class ByteDanceMediationAdapter
                     {
                         // Pangle's image resource comes in the form of a URL which needs to be fetched in a non-blocking manner
                         log( "Adding native ad icon (" + nativeAd.getIcon().getImageUrl() + ") to queue to be fetched" );
-                        iconDrawableFuture = createDrawableFuture( nativeAd.getIcon().getImageUrl(), activity.getResources() );
+                        iconDrawableFuture = createDrawableFuture( nativeAd.getIcon().getImageUrl(), context.getResources() );
                     }
 
                     // Pangle's media view can be either a video or image (which they don't provide a view for)
@@ -1131,7 +1138,7 @@ public class ByteDanceMediationAdapter
                         {
                             // Pangle's image resource comes in the form of a URL which needs to be fetched in a non-blocking manner
                             log( "Adding native ad media (" + ttMediaImage.getImageUrl() + ") to queue to be fetched" );
-                            imageDrawableFuture = createDrawableFuture( ttMediaImage.getImageUrl(), activity.getResources() );
+                            imageDrawableFuture = createDrawableFuture( ttMediaImage.getImageUrl(), context.getResources() );
                         }
                     }
 
@@ -1172,7 +1179,7 @@ public class ByteDanceMediationAdapter
                             }
                             else if ( finalMediaViewImageDrawable != null )
                             {
-                                mediaView = new ImageView( activity );
+                                mediaView = new ImageView( context );
                                 ( (ImageView) mediaView ).setImageDrawable( finalMediaViewImageDrawable );
                             }
                             else